Correccion de cambios
